Cleanup file naming and organization
